feat: Complete Docker containerization and critical bug fixes
Major changes:
- Docker containerization with PostgreSQL and Redis
- Migration from SQLite to PostgreSQL with data integrity
- Fixed critical bugs in request handling (user_id vs telegram_id)
- Fixed Aiogram 3.x middleware registration
- Fixed validation for request descriptions (10 chars limit)
- Allow managers/admins to manage their own requests
- Added all 4 roles to admin user
- Fixed pending user button visibility
- Added hot-reload development setup
- Updated Memory Bank with all completed tasks

Technical improvements:
- BIGINT for telegram_id in PostgreSQL
- Proper foreign key mapping in migration
- Fixed request display and creation
- Removed unnecessary shift-related messages
- Enhanced role switching in profile
- Production-ready Docker configuration

Files added/modified:
- Docker configuration files (Dockerfile, docker-compose*.yml)
- Database scripts (init_postgres.sql, migration script)
- Core application files (handlers, services, models)
- Documentation (DOCKER_SETUP.md, QUICK_DEV_START.md)
- Memory Bank updates (activeContext.md, tasks.md)

Status: Production-ready with development hot-reload support

+++ uk_management_bot/database/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, Text
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean, Text
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 from database.session import Base
@@ -7,7 +7,7 @@ class User(Base):
     __tablename__ = "users"
     
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, index=True, nullable=False)
+    telegram_id = Column(BigInteger, unique=True, index=True, nullable=False)
     username = Column(String(255), nullable=True)
     first_name = Column(String(255), nullable=True)
     last_name = Column(String(255), nullable=True)
